Rename transformation modules with rewrite prefix
