+++ src/transform/mod.rs
@@ -2,10 +2,10 @@ pub(crate) mod assert;
 pub(crate) mod class_def;
 pub(crate) mod decorator;
 pub(crate) mod expr;
-pub(crate) mod for_loop;
+pub(crate) mod rewrite_for_loop;
 pub(crate) mod gen;
 pub(crate) mod import;
-pub(crate) mod match_case;
+pub(crate) mod rewrite_match_case;
 pub(crate) mod truthy;
-pub(crate) mod try_except;
+pub(crate) mod rewrite_try_except;
 pub(crate) mod with;
